feat: Add alternative websocket into MRe
The default MRe SDK consumes all websocket cnnections, assuming
they're all from AltSpaceVR. That's not going to be the case for us: we
want a webapp interface to bridge communications from the outside
with AltSpace via a browser.

As a PoC of this, a status button has been added to display the
rooms+users connected to the MRe.