feat: MRe webapp is now a buttplug client
This change is a mess, but what's new for a PoC.
The jist is that we've borrowed a bunch of logic from the
https://github.com/qdot/simple-teledildonics-app example
to make the MRe's web-interface a buttplug client.

Each visit to the web-app will create a websocket to first
allocate a "room" ID (bad terminology; I'll come up with
something better later). This is presented on the page as a
6 character access code. This is used when the buttplug client
in the webapp creates its connection to the MRe server.

The 6 character access code also the value that the user is prompted
for when they connect to the world hosting the MRe. This acts as a
link between the web-app buttplug connection and the VR hosted
interface.

Clicking the buzz button (only visible to the user with the valid
connection) will create a buttplug client in the MRe server to connect
to the buttplug server that's connected to the buttplug client in the
browser and sends a vibration command for 30 seconds.